씬 체인지 오류 수정

+++ 2023_winapi_framework/DamageCaster.cpp
@@ -15,7 +15,7 @@ DamageCaster::~DamageCaster()
 	OnDamageOverCallback = nullptr;
 }
 
-void DamageCaster::OnDamage(float _damage)
+void DamageCaster::OnDamage(int _damage)
 {
 	m_currentHealth -= _damage;
 	if (m_currentHealth <= 0.f) {
